Improve userfunctionality-mgt service component

+++ components/user-functionality-mgt/org.wso2.carbon.identity.user.functionality.mgt/src/main/java/org/wso2/carbon/identity/user/functionality/mgt/UserFunctionalityManagerImpl.java
@@ -19,14 +19,25 @@
 package org.wso2.carbon.identity.user.functionality.mgt;
 
 import org.apache.commons.collections.MapUtils;
+import org.apache.commons.lang.StringUtils;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.wso2.carbon.identity.core.util.IdentityTenantUtil;
 import org.wso2.carbon.identity.core.util.IdentityUtil;
 import org.wso2.carbon.identity.user.functionality.mgt.dao.UserFunctionalityManagerDAO;
 import org.wso2.carbon.identity.user.functionality.mgt.dao.UserFunctionalityPropertyDAO;
 import org.wso2.carbon.identity.user.functionality.mgt.dao.impl.UserFunctionalityManagerDAOImpl;
 import org.wso2.carbon.identity.user.functionality.mgt.dao.impl.UserFunctionalityPropertyDAOImpl;
+import org.wso2.carbon.identity.user.functionality.mgt.exception.UserFunctionalityManagementClientException;
 import org.wso2.carbon.identity.user.functionality.mgt.exception.UserFunctionalityManagementException;
 import org.wso2.carbon.identity.user.functionality.mgt.exception.UserFunctionalityManagementServerException;
+import org.wso2.carbon.identity.user.functionality.mgt.internal.UserFunctionalityManagerComponentDataHolder;
 import org.wso2.carbon.identity.user.functionality.mgt.model.FunctionalityLockStatus;
+import org.wso2.carbon.user.api.UserStoreException;
+import org.wso2.carbon.user.api.UserStoreManager;
+import org.wso2.carbon.user.core.UniqueIDUserStoreManager;
+import org.wso2.carbon.user.core.constants.UserCoreErrorConstants;
+import org.wso2.carbon.user.core.service.RealmService;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -39,17 +50,26 @@ public class UserFunctionalityManagerImpl implements UserFunctionalityManager {
 
     private UserFunctionalityManagerDAO userFunctionalityManagerDAO = new UserFunctionalityManagerDAOImpl();
     private UserFunctionalityPropertyDAO userFunctionalityPropertyDAO = new UserFunctionalityPropertyDAOImpl();
+    private static final Log log = LogFactory.getLog(UserFunctionalityManagerImpl.class);
 
     /**
      * {@inheritDoc}
      */
     @Override
     public FunctionalityLockStatus getLockStatus(String userId, int tenantId, String functionalityIdentifier)
-            throws UserFunctionalityManagementServerException {
+            throws UserFunctionalityManagementException {
 
         if (!isPerUserFunctionalityLockingEnabled()) {
             throw new UnsupportedOperationException("Per-user functionality locking is not enabled.");
         }
+
+        if (StringUtils.isEmpty(userId) || !isUserIdExists(userId, tenantId)) {
+            if (log.isDebugEnabled()) {
+                log.debug("Cannot retrieve user from userId: " + userId);
+            }
+            throw buildUserNotFoundError();
+        }
+
         FunctionalityLockStatus
                 functionalityLockStatus =
                 userFunctionalityManagerDAO.getFunctionalityLockStatus(userId, tenantId, functionalityIdentifier);
@@ -69,10 +89,16 @@ public class UserFunctionalityManagerImpl implements UserFunctionalityManager {
      */
     @Override
     public Map<String, String> getProperties(String userId, int tenantId, String functionalityIdentifier)
-            throws UserFunctionalityManagementServerException {
+            throws UserFunctionalityManagementException {
 
         if (!isPerUserFunctionalityLockingEnabled()) {
             throw new UnsupportedOperationException("Per-user functionality locking is not enabled.");
+        }
+        if (StringUtils.isEmpty(userId) || !isUserIdExists(userId, tenantId)) {
+            if (log.isDebugEnabled()) {
+                log.debug("Cannot retrieve user from userId: " + userId);
+            }
+            throw buildUserNotFoundError();
         }
         return userFunctionalityPropertyDAO.getAllProperties(userId, tenantId, functionalityIdentifier);
     }
@@ -83,11 +109,18 @@ public class UserFunctionalityManagerImpl implements UserFunctionalityManager {
     @Override
     public void setProperties(String userId, int tenantId, String functionalityIdentifier,
                               Map<String, String> functionalityLockProperties)
-            throws UserFunctionalityManagementServerException {
+            throws UserFunctionalityManagementException {
 
         if (!isPerUserFunctionalityLockingEnabled()) {
             throw new UnsupportedOperationException("Per-user functionality locking is not enabled.");
         }
+        if (StringUtils.isEmpty(userId) || !isUserIdExists(userId, tenantId)) {
+            if (log.isDebugEnabled()) {
+                log.debug("Cannot retrieve user from userId: " + userId);
+            }
+            throw buildUserNotFoundError();
+        }
+
         Map<String, String> existingProperties = getProperties(userId, tenantId, functionalityIdentifier);
         if (MapUtils.isNotEmpty(functionalityLockProperties)) {
             addOrUpdateProperties(functionalityLockProperties, existingProperties, userId, tenantId,
@@ -106,11 +139,16 @@ public class UserFunctionalityManagerImpl implements UserFunctionalityManager {
         if (!isPerUserFunctionalityLockingEnabled()) {
             throw new UnsupportedOperationException("Per-user functionality locking is not enabled.");
         }
+        if (StringUtils.isEmpty(userId) || !isUserIdExists(userId, tenantId)) {
+            if (log.isDebugEnabled()) {
+                log.debug("Cannot retrieve user from userId: " + userId);
+            }
+            throw buildUserNotFoundError();
+        }
         long unlockTime = Long.MAX_VALUE;
         if (timeToLock >= 0) {
             unlockTime = System.currentTimeMillis() + timeToLock;
         }
-
         FunctionalityLockStatus functionalityLockStatus =
                 userFunctionalityManagerDAO.getFunctionalityLockStatus(userId, tenantId, functionalityIdentifier);
         if (functionalityLockStatus != null) {
@@ -144,10 +182,16 @@ public class UserFunctionalityManagerImpl implements UserFunctionalityManager {
      */
     @Override
     public void unlock(String userId, int tenantId, String functionalityIdentifier)
-            throws UserFunctionalityManagementServerException {
+            throws UserFunctionalityManagementException {
 
         if (!isPerUserFunctionalityLockingEnabled()) {
             throw new UnsupportedOperationException("Per-user functionality locking is not enabled.");
+        }
+        if (StringUtils.isEmpty(userId) || !isUserIdExists(userId, tenantId)) {
+            if (log.isDebugEnabled()) {
+                log.debug("Cannot retrieve user from userId: " + userId);
+            }
+            throw buildUserNotFoundError();
         }
         userFunctionalityManagerDAO.deleteMappingForUser(userId, tenantId, functionalityIdentifier);
     }
@@ -157,10 +201,16 @@ public class UserFunctionalityManagerImpl implements UserFunctionalityManager {
      */
     @Override
     public void deleteAllPropertiesForUser(String userId, int tenantId, String functionalityIdentifier)
-            throws UserFunctionalityManagementServerException {
+            throws UserFunctionalityManagementException {
 
         if (!isPerUserFunctionalityLockingEnabled()) {
             throw new UnsupportedOperationException("Per-user functionality locking is not enabled.");
+        }
+        if (StringUtils.isEmpty(userId) || !isUserIdExists(userId, tenantId)) {
+            if (log.isDebugEnabled()) {
+                log.debug("Cannot retrieve user from userId: " + userId);
+            }
+            throw buildUserNotFoundError();
         }
         userFunctionalityPropertyDAO.deleteAllPropertiesForUser(userId, tenantId, functionalityIdentifier);
 
@@ -172,10 +222,16 @@ public class UserFunctionalityManagerImpl implements UserFunctionalityManager {
     @Override
     public void deletePropertiesForUser(String userId, int tenantId, String functionalityIdentifier,
                                         Set<String> propertiesToDelete)
-            throws UserFunctionalityManagementServerException {
+            throws UserFunctionalityManagementException {
 
         if (!isPerUserFunctionalityLockingEnabled()) {
             throw new UnsupportedOperationException("Per-user functionality locking is not enabled.");
+        }
+        if (StringUtils.isEmpty(userId) || !isUserIdExists(userId, tenantId)) {
+            if (log.isDebugEnabled()) {
+                log.debug("Cannot retrieve user from userId: " + userId);
+            }
+            throw buildUserNotFoundError();
         }
         userFunctionalityPropertyDAO.deletePropertiesForUser(userId, tenantId, functionalityIdentifier, propertiesToDelete);
 
@@ -230,5 +286,65 @@ public class UserFunctionalityManagerImpl implements UserFunctionalityManager {
 
         return Boolean.parseBoolean(
                 IdentityUtil.getProperty(UserFunctionalityMgtConstants.ENABLE_PER_USER_FUNCTIONALITY_LOCKING));
+    }
+
+    private boolean isUserIdExists(String userId, int tenantId) throws UserFunctionalityManagementClientException,
+            UserFunctionalityManagementServerException {
+
+        boolean isUserExists;
+
+        try {
+            UniqueIDUserStoreManager uniqueIdEnabledUserStoreManager = getUniqueIdEnabledUserStoreManager(
+                    UserFunctionalityManagerComponentDataHolder.getInstance().getRealmService(),
+                    IdentityTenantUtil.getTenantDomain(tenantId));
+
+            isUserExists = uniqueIdEnabledUserStoreManager.isExistingUserWithID(userId);
+            return isUserExists;
+        } catch (UserStoreException e) {
+            if (isUserNotExistingError(e, userId)) {
+                if (log.isDebugEnabled()) {
+                    log.debug("Cannot retrieve user from userId: " + userId, e);
+                }
+                throw buildUserNotFoundError();
+            }
+            throw new UserFunctionalityManagementServerException(
+                    UserFunctionalityMgtConstants.ErrorMessages.ERROR_OCCURRED_WHILE_RETRIEVING_USER.getCode(),
+                    UserFunctionalityMgtConstants.ErrorMessages.ERROR_OCCURRED_WHILE_RETRIEVING_USER.getDescription());
+        }
+
+    }
+
+    private UniqueIDUserStoreManager getUniqueIdEnabledUserStoreManager(RealmService realmService, String tenantDomain)
+            throws UserStoreException, UserFunctionalityManagementClientException {
+
+        UserStoreManager userStoreManager = realmService.getTenantUserRealm(
+                IdentityTenantUtil.getTenantId(tenantDomain)).getUserStoreManager();
+        if (!(userStoreManager instanceof UniqueIDUserStoreManager)) {
+            if (log.isDebugEnabled()) {
+                String msg = "Provided user store manager does not support unique user IDs in the tenant domain" +
+                        tenantDomain;
+                log.debug(msg);
+            }
+            throw buildUserNotFoundError();
+        }
+        return (UniqueIDUserStoreManager) userStoreManager;
+    }
+
+    private boolean isUserNotExistingError(UserStoreException e, String userId) {
+
+        if (log.isDebugEnabled()) {
+            String msg = "Provided user corresponding to the userid" + userId + "does not exist";
+            log.debug(msg);
+        }
+        return e instanceof org.wso2.carbon.user.core.UserStoreException &&
+                UserCoreErrorConstants.ErrorMessages.ERROR_CODE_NON_EXISTING_USER.getCode().equals(
+                        ((org.wso2.carbon.user.core.UserStoreException) e).getErrorCode());
+    }
+
+    private UserFunctionalityManagementClientException buildUserNotFoundError() {
+
+        return new UserFunctionalityManagementClientException(
+                UserFunctionalityMgtConstants.ErrorMessages.USER_NOT_FOUND.getCode(),
+                UserFunctionalityMgtConstants.ErrorMessages.USER_NOT_FOUND.getDescription());
     }
 }
